Allow the coveralls job to fail
In the last green Travis build[1], coveralls failed with "command
not found". We fixed that particular error in this build but there
are still missing environment variables. Since the scope of this
change is to replace the existing Travis behavior, we'll stop here
and let the coveralls fail but ignore that fact.

[1]:https://travis-ci.org/github/rollbar/rollbar-php-laravel/jobs/743942091