feat: add test matrix shell
This is the basic operation: stand up the framework, configure
Rollbar within that framework, then invoke a log message. Rig
Rollbar so that logged messages get written to a file, that we
then check exists and contains the desired value.

Later iterations will add the specifics for each framework-version,
which may require multiple action files: the setup varies slightly
between framework and framework-version.