refac: break out into lineage branches
Now that we've proved out the support (from prior runs where the
matrix encompassed all supported versions and was hitting released
tags) and we have a baseline of known support, we can break out
the code into lineage branches. In this regime (which matches
rollbar-php operation), commits to the lineage branch are checked
against the documented support for that lineage.